vo_gpu: check for HDR peak detection in dumb mode too
Similar spirit to ce0c4ae983f8. check_gl_features() has a confusing
early-return. This also adds compute_hdr_peak to the list of options
that is copied to the dumb-mode options struct, since it seems to make a
difference. Otherwise it would be impossible to disable HDR peak
detection in dumb mode.