cocoa-cb: fix usage of wrong fbo when off-screen
when we transitioned to the new libmpv API with commit 0861f9f i
reintroduced an old bug that was fixed with 38d6346 because of a dumb
rebasing error on my part. the branch i based the libmpv changed on was
originally without the fbo fix and on rebasing i forgot to change the
variable to the proper one, basically deactivating the fix.