js: correctness: use integer range checks
There were some cases where a js number (double) was blindly casted to
int or uint64, but that can be undefined behavior (out of range to int)
or wrong (negative to uint).

Now the code throws a js error if the value is out of range.

Additionally, commit e3642f1a added these checks for the new hooks API,
but incorrectly tested int64 range rather than uint64. Fix this too.